RMC-45 Do additional tests for RemindersEndpoints and improve endPoints and PetService

+++ backend/src/main/java/com/rmc/pejo/endpoints/PetEndPoint.java
@@ -2,23 +2,23 @@ package com.rmc.pejo.endpoints;
 
 import com.rmc.pejo.entity.Pet;
 import com.rmc.pejo.entity.Reminder;
+import com.rmc.pejo.exceptions.ResourceNotFoundException;
 import com.rmc.pejo.service.PetService;
+import com.rmc.pejo.service.ReminderService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import java.util.List;
-import java.util.Optional;
 import java.util.Set;
 
 @CrossOrigin
 @RestController
 @RequestMapping("pets")
+@RequiredArgsConstructor
 public class PetEndPoint {
     private final PetService service;
-
-    public PetEndPoint(PetService petService) {
-        this.service = petService;
-    }
+    private final ReminderService reminderService;
 
     @PostMapping
     public Pet save(@Valid @RequestBody Pet pet) {
@@ -31,10 +31,9 @@ public class PetEndPoint {
     }
 
     @GetMapping("{id}")
-    public Optional<Pet> get(@PathVariable Long id) {
-        return service.get(id);
+    public Pet get(@PathVariable Long id) {
+        return service.get(id).orElseThrow(ResourceNotFoundException::new);
     }
-
 
     @PutMapping
     public Pet update(@Valid @RequestBody Pet pet) {
@@ -42,18 +41,20 @@ public class PetEndPoint {
     }
 
     @PutMapping("{id}")
-    public void addReminder(@PathVariable Long id, @RequestBody Reminder reminder) {
-        service.addReminder(id, reminder);
+    public Pet addReminder(@PathVariable Long id, @RequestBody Reminder reminder) {
+        return service.addReminder(id, reminder).orElseThrow(ResourceNotFoundException::new);
     }
 
     @DeleteMapping("{id}")
     public String deleteById(@PathVariable("id") Long id) {
+        if (service.get(id).isEmpty()) throw new ResourceNotFoundException();
         service.delete(id);
         return "Deleted successfully";
     }
 
     @GetMapping("reminder/{reminderId}")
     public Set<Pet> getPetsByReminderId(@PathVariable("reminderId") Long reminderId) {
+        if (reminderService.get(reminderId).isEmpty()) throw new ResourceNotFoundException();
         return service.getPetsByReminderId(reminderId);
     }
 }
